finished speed extractions
